requirements.txt
python-daemon
pyev
RPi.GPIO

